refactor: Blinds
